requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.23.2
pydantic==2.5.2
python-dotenv==1.0.0
pytest==8.2.0
pytest-cov==4.1.0
psycopg[binary]
#pip install -r requirements.txt
